Update TimePlayed Bug in Statistics (ii)

+++ src/code/3_states/5_statistics/statistics.cpp
@@ -62,11 +62,11 @@ void Statistics::drawTimePlayed(int& posY){
     posY += textSize + 5;
 
     DrawText("Total Time Played",                          posX, posY, textSize, RAYWHITE);
-    DrawText(TextFormat("%dm %ds", (stats.totalTimePlayed / 60.0),         ((int)stats.totalTimePlayed % 60)),      posX + offset, posY, textSize, LIME);
+    DrawText(TextFormat("%dm %ds", ((int)stats.totalTimePlayed / 60),      ((int)stats.totalTimePlayed % 60)),      posX + offset, posY, textSize, LIME);
     posY += textSize + 5;
 
     DrawText("Longest Game",                               posX, posY, textSize, RAYWHITE);
-    DrawText(TextFormat("%dm %ds", (stats.maxTimePlayedPerGame / 60.0),    ((int)stats.maxTimePlayedPerGame % 60)), posX + offset, posY, textSize, LIME);
+    DrawText(TextFormat("%dm %ds", ((int)stats.maxTimePlayedPerGame / 60), ((int)stats.maxTimePlayedPerGame % 60)), posX + offset, posY, textSize, LIME);
     posY += textSize + 5;
 
     DrawText("Avg Time / Game",                            posX, posY, textSize, RAYWHITE);
